utils: move tmpbuf_file to test_utils
