+++ src/v/storage/tests/offset_index_utils_tests.cc
@@ -10,8 +10,8 @@
 #include "serde/serde.h"
 #include "storage/segment_index.h"
 #include "test_utils/fixture.h"
+#include "test_utils/tmpbuf_file.h"
 #include "utils/file_io.h"
-#include "utils/tmpbuf_file.h"
 
 #include <seastar/core/seastar.hh>
 #include <seastar/core/shared_ptr.hh>
